ICU-22261 Refactor error handling for internal MF2 array-allocating functions
Several functions used to copy between vectors and arrays were written
not to take a UErrorCode, but rather to take an int argument by reference
and set it to 0 to indicate a memory allocation error. To be more
consistent with ICU style, change these functions to take a UErrorCode
argument.

Several data model classes (Reserved, Matcher and Pattern) are also
modified to add a `bogus` flag to indicate that a copy failed. Previously,
the "bogus" state was denoted by setting an internal length field to 0,
but that was wrong because a valid instance of these classes can have
length 0 and that state should be distinguished from an uninitialized
state.